requirements.txt
pyspark==3.1.2
apache-airflow==2.6.0
pandas==1.3.4
pyarrow==10.0.1
zipfile36==0.1.3
kaggle==1.5.13
jupyterlab==3.1.12
requests
apache-airflow-providers-apache-spark==2.1.0
scikit-learn==1.0.1
joblib==1.1.0
fastparquet
dask